Exceptionhandling in BookService hinzugefügt

+++ 3.3/src/main/java/edu/hm/ba/classic/services/BookServiceImpl.java
@@ -2,12 +2,17 @@ package edu.hm.ba.classic.services;
 
 import edu.hm.ba.classic.entities.Book;
 import edu.hm.ba.classic.entities.User;
+import edu.hm.ba.classic.exception.BookDoesNotExistException;
+import edu.hm.ba.classic.exception.CouldNotCreateBookException;
+import edu.hm.ba.classic.exception.DuplicateBookException;
 import edu.hm.ba.classic.persistence.BookRepository;
 import edu.hm.ba.classic.persistence.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Optional;
 
 /**
  * Implementation of the BookService.
@@ -15,6 +20,11 @@ import java.util.Collection;
  */
 @Service
 public class BookServiceImpl implements BookService {
+
+    /**
+     * Length of an isbn code.
+     */
+    public static final int ISBN_LENGTH = 13;
 
     /**
      * Book Repository.
@@ -34,22 +44,43 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void addBook(Book book) {
-        bookRepository.save(book);
+    public Book addBook(Book book) {
+        if (!checkIsbn(book.getIsbn()) || book.getTitle().isEmpty() || book.getAuthor().isEmpty()) {
+            throw new CouldNotCreateBookException("Unable to add book with isbn " + book.getIsbn() +
+                    " title " + book.getTitle() + " and author " + book.getAuthor());
+        }
+        book.setIsbn(book.getIsbn().replace("-", "")); //mögliche Trennzeichen entfernen
+        //Auf Duplikat prüfen
+        for (Book b : getBooks()) {
+            if (b.getIsbn().equals(book.getIsbn())) {
+                throw new DuplicateBookException("The book with isbn " + book.getIsbn() + " does already exist.");
+            }
+        }
+        return bookRepository.save(book);
     }
 
     @Override
-    public void deleteBook(String isbn) {
+    public Book deleteBook(String isbn) {
+        Optional<Book> toDelete = bookRepository.findById(isbn);
+        if (toDelete == null) {
+            throw new BookDoesNotExistException("The book with isbn " + isbn + " can not be found.");
+        }
+        //bookRepository.delete(toDelete);
         bookRepository.deleteById(isbn);
+        return toDelete.orElse(null);
     }
 
     @Override
-    public void updateBook(Book book) {
-        Book toUpdate = bookRepository.getOne(book.getIsbn());
+    public Book updateBook(String isbn, Book book) {
+        Book toUpdate = bookRepository.getOne(isbn);
+        if (!isbn.equals(book.getIsbn()) || book.getTitle().isEmpty() || book.getAuthor().isEmpty()) {
+            throw new CouldNotCreateBookException("Unable to update book with isbn " + book.getIsbn() +
+                    " title " + book.getTitle() + " and author " + book.getAuthor());
+        }
         toUpdate.setTitle(book.getTitle());
         toUpdate.setAuthor(book.getAuthor());
         toUpdate.setCategory(book.getCategory());
-        bookRepository.save(toUpdate);
+        return bookRepository.save(toUpdate);
     }
 
     @Override
@@ -72,5 +103,40 @@ public class BookServiceImpl implements BookService {
                 bookRepository.save(toReturn);
             }
         }
+    }
+
+    /**
+     * Checks if there is a valid isbn number.
+     * @param isbn isbn to check
+     * @return true if the isbn is valid
+     */
+    private boolean checkIsbn(String isbn) {
+        if (isbn.isEmpty()) {
+            return false;
+        } else {
+            isbn = isbn.replace("-", ""); //mögliche Trennzeichen entfernen
+        }
+        if (isbn.length() != ISBN_LENGTH) {
+            return false;
+        }
+        ArrayList<Integer> ints = convertToList(isbn);
+        int checksum = (ints.get(0) + ints.get(2) + ints.get(4) + ints.get(6) + ints.get(8) + ints.get(10) + ints.get(12) + 3 * (ints.get(1) + ints.get(3) + ints.get(5) + ints.get(7) + ints.get(9) + ints.get(11))) % 10;
+        if (checksum == 0) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Converts a String of numbers to a list of integers.
+     * @param isbn String to convert
+     * @return list of integers
+     */
+    private ArrayList<Integer> convertToList(String isbn) {
+        ArrayList<Integer> result = new ArrayList<>();
+        for (char c: isbn.toCharArray()) {
+            result.add(Character.getNumericValue(c));
+        }
+        return result;
     }
 }
